# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  github_com_basilex_promenade_pkg_response.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_basilex_promenade_pkg_response.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      pages:
        type: integer
      total:
        type: integer
    type: object
  github_com_basilex_promenade_pkg_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Error'
      message:
        type: string
      pagination:
        $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Pagination'
      status:
        type: string
    type: object
...
  internal_contexts_warehouse_product_adapter_http.CreateProductRequest:
    properties:
      name:
        example: Laptop Dell XPS 15
        type: string
      sku:
        example: PROD-001
        type: string
    required:
    - name
    - sku
    type: object
...
  internal_contexts_warehouse_product_adapter_http.ProductListResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      products:
        items:
          $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
        type: array
      total:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  internal_contexts_warehouse_product_adapter_http.UpdateInventorySettingsRequest:
    properties:
      allow_backorder:
        example: false
        type: boolean
      track_inventory:
        example: true
        type: boolean
    type: object
  internal_contexts_warehouse_product_adapter_http.UpdateProductRequest:
    properties:
      brand:
        example: Dell
        type: string
      category:
        example: Electronics
        type: string
      description:
        example: High-performance laptop
        type: string
      name:
        example: Laptop Dell XPS 15
        type: string
    required:
    - name
    type: object
...
host: localhost:8081
info:
  contact:
    email: alexander.vasilenko@gmail.com
    name: API Support
    url: https://github.com/basilex/promenade
  description: Modern backend platform for customer management, orders, and business
    workflows with clean DDD architecture
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Promenade Platform
  version: 0.1.0
paths:
...
  /warehouse/products:
    get:
      consumes:
      - application/json
      description: Get paginated list of products
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductListResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: Create a new product with SKU and name
      parameters:
      - description: Product creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Create product
      tags:
      - Products
  /warehouse/products/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete a product by ID
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Delete product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: Get product details by product ID
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Get product by ID
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: Update product details
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Product update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update product
      tags:
      - Products
  /warehouse/products/{id}/activate:
    post:
      consumes:
      - application/json
      description: Transition product to Active status
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Activate product
      tags:
      - Products
  /warehouse/products/{id}/deactivate:
    post:
      consumes:
      - application/json
      description: Transition product to OutOfStock status
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Deactivate product
      tags:
      - Products
  /warehouse/products/{id}/discontinue:
    post:
      consumes:
      - application/json
      description: Mark product as discontinued
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Discontinue product
      tags:
      - Products
  /warehouse/products/{id}/inventory-settings:
    put:
      consumes:
      - application/json
      description: Update inventory tracking and backorder settings
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Inventory settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.UpdateInventorySettingsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Update inventory settings
      tags:
      - Products
  /warehouse/products/{id}/physical-properties:
    put:
      consumes:
      - application/json
      description: Set weight and dimensions for physical product
      parameters:
      - description: Product ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Physical properties
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.SetPhysicalPropertiesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/internal_contexts_warehouse_product_adapter_http.ProductResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_basilex_promenade_pkg_response.Response'
      summary: Set physical properties
      tags:
      - Products
...
schemes:
- http
- https
swagger: "2.0"